Make data package
