+++ src/main/java/com/titaniumtigers4829/TigerHelpers.java
@@ -1,4 +1,4 @@
-// TigerHelpers v1.03 (REQUIRES LLOS 2025.0 OR LATER)
+// TigerHelpers v1.05 (REQUIRES LLOS 2025.0 OR LATER)
 
 package com.titaniumtigers4829;
 
@@ -14,9 +14,12 @@ import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.networktables.TimestampedDoubleArray;
-import java.util.Arrays;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+
+import com.titaniumtigers4829.data.IMUData;
+import com.titaniumtigers4829.data.PoseEstimate;
+import com.titaniumtigers4829.data.RawFiducial;
 
 /**
  * TigerHelpers is built on top of LimelightHelpers, providing a set of static methods and classes
@@ -29,299 +32,6 @@ import java.util.concurrent.ConcurrentHashMap;
 public class TigerHelpers {
 
   private static final Map<String, DoubleArrayEntry> doubleArrayEntries = new ConcurrentHashMap<>();
-
-  /** Represents a Limelight Raw Fiducial result from Limelight's NetworkTables output. */
-  public static class RawFiducial {
-    /** The id of the april tag */
-    public int id = 0;
-
-    /**
-     * The horizontal offset of the target from the Limelight's principle pixel (the images central
-     * pixel) in degrees. Positive values mean the target is to the right of the crosshair.
-     */
-    public double txnc = 0;
-
-    /**
-     * The vertical offset of the target from the Limelight's principle pixel (the images central
-     * pixel) in degrees. Positive values mean the target is below the crosshair.
-     */
-    public double tync = 0;
-
-    /**
-     * The area of the target in the Limelight's view as a percentage of the total image area. So,
-     * if the target takes up 10% of the image, this value will be 0.1.
-     */
-    public double ta = 0;
-
-    /** The distance from the Limelight to the april tag in meters. */
-    public double distToCamera = 0;
-
-    /** The distance from the robot to the april tag in meters. */
-    public double distToRobot = 0;
-
-    /**
-     * The ambiguity of the april tag. This ranges from 0 to 1, a lower values meaning less
-     * ambiguous. The higher this is, the more likely it is that you will see ambiguity flipping.
-     * Ambiguity flipping is when there are multiple possible "solutions" for the pose estimate, so
-     * if this is higher your estimate will be less reliable. The best way to reduce this it to make
-     * the april tag look as trapezoidal as possible from the Limelight's perspective.
-     */
-    public double ambiguity = 0;
-
-    /**
-     * Initializes a RawFidicual with the provided values. This represents data from reading a
-     * single april tag.
-     *
-     * @param id The id of the april tag
-     * @param txnc The horizontal offset of the target from the Limelight's principle pixel (the
-     *     images central pixel) in degrees.
-     * @param tync The vertical offset of the target from the Limelight's principle pixel (the
-     *     images central pixel) in degrees.
-     * @param ta The area of the target in the Limelight's view as a percentage of the total image
-     *     area.
-     * @param distToCamera The distance from the Limelight to the april tag in meters.
-     * @param distToRobot The distance from the robot to the april tag in meters.
-     * @param ambiguity The ambiguity of the april tag. This ranges from 0 to 1, a lower values
-     *     meaning less ambiguous.
-     */
-    public RawFiducial(
-        int id,
-        double txnc,
-        double tync,
-        double ta,
-        double distToCamera,
-        double distToRobot,
-        double ambiguity) {
-      this.id = id;
-      this.txnc = txnc;
-      this.tync = tync;
-      this.ta = ta;
-      this.distToCamera = distToCamera;
-      this.distToRobot = distToRobot;
-      this.ambiguity = ambiguity;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      if (this == obj) return true;
-      if (obj == null || getClass() != obj.getClass()) return false;
-      RawFiducial other = (RawFiducial) obj;
-      return id == other.id
-          && Double.compare(txnc, other.txnc) == 0
-          && Double.compare(tync, other.tync) == 0
-          && Double.compare(ta, other.ta) == 0
-          && Double.compare(distToCamera, other.distToCamera) == 0
-          && Double.compare(distToRobot, other.distToRobot) == 0
-          && Double.compare(ambiguity, other.ambiguity) == 0;
-    }
-  }
-
-  /**
-   * Represents a Limelight pose estimate from Limelight's NetworkTables output. This is for any
-   * botpose estimate, so MegaTag1 or MegaTag2
-   */
-  public static class PoseEstimate {
-    /** The estimated 2D pose of the robot, including position and rotation. */
-    public Pose2d pose;
-
-    /** The timestamp of the pose estimate in seconds since the Limelight booted up. */
-    public double timestampSeconds;
-
-    /** The latency of the pose estimate in milliseconds. */
-    public double latency;
-
-    /** The number of april tags used to calculate the pose estimate. */
-    public int tagCount;
-
-    /** The max distance, in meters, between april tags used to calculate the pose estimate. */
-    public double tagSpan;
-
-    /**
-     * The average distance, in meters, between the Limelight and the april tags used to calculate
-     * the pose estimate.
-     */
-    public double avgTagDist;
-
-    /** The average area, in square meters, of april tags used to calculate the pose estimate. */
-    public double avgTagArea;
-
-    /** An array of RawFiducial used to calculate the pose estimate. */
-    public RawFiducial[] rawFiducials;
-
-    /** true if the pose estimate is calculated using MegaTag2, false if using MegaTag1. */
-    public boolean isMegaTag2;
-
-    /** Initializes an "empty" PoseEstimate object with default values */
-    public PoseEstimate() {
-      this.pose = new Pose2d();
-      this.timestampSeconds = 0;
-      this.latency = 0;
-      this.tagCount = 0;
-      this.tagSpan = 0;
-      this.avgTagDist = 0;
-      this.avgTagArea = 0;
-      this.rawFiducials = new RawFiducial[] {};
-      this.isMegaTag2 = false;
-    }
-
-    /**
-     * Initializes a PoseEstimate object with the provided data.
-     *
-     * @param pose The estimated 2D pose of the robot, including position and rotation.
-     * @param timestampSeconds The timestamp of the pose estimate in seconds since the Limelight
-     *     booted up.
-     * @param latency The latency of the pose estimate in milliseconds.
-     * @param tagCount The number of april tags used to calculate the pose estimate.
-     * @param tagSpan The max distance, in meters, between april tags used to calculate the pose
-     *     estimate.
-     * @param avgTagDist The average distance, in meters, between the Limelight and the april tags
-     *     used to calculate the pose estimate.
-     * @param avgTagArea The average area, in square meters, of april tags used to calculate the
-     *     pose estimate.
-     * @param rawFiducials An array of RawFiducial used to calculate the pose estimate.
-     * @param isMegaTag2 true if the pose estimate is calculated using MegaTag2, false if using
-     *     MegaTag1.
-     */
-    public PoseEstimate(
-        Pose2d pose,
-        double timestampSeconds,
-        double latency,
-        int tagCount,
-        double tagSpan,
-        double avgTagDist,
-        double avgTagArea,
-        RawFiducial[] rawFiducials,
-        boolean isMegaTag2) {
-
-      this.pose = pose;
-      this.timestampSeconds = timestampSeconds;
-      this.latency = latency;
-      this.tagCount = tagCount;
-      this.tagSpan = tagSpan;
-      this.avgTagDist = avgTagDist;
-      this.avgTagArea = avgTagArea;
-      this.rawFiducials = rawFiducials;
-      this.isMegaTag2 = isMegaTag2;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      if (this == obj) return true;
-      if (obj == null || getClass() != obj.getClass()) return false;
-      PoseEstimate that = (PoseEstimate) obj;
-      // We don't compare the timestampSeconds as it isn't relevant for equality and makes
-      // unit testing harder
-      return Double.compare(that.latency, latency) == 0
-          && tagCount == that.tagCount
-          && Double.compare(that.tagSpan, tagSpan) == 0
-          && Double.compare(that.avgTagDist, avgTagDist) == 0
-          && Double.compare(that.avgTagArea, avgTagArea) == 0
-          && pose.equals(that.pose)
-          && Arrays.equals(rawFiducials, that.rawFiducials);
-    }
-
-    @Override
-    public String toString() {
-      // Convert rawFiducials to a string of just their IDs
-      StringBuilder fiducialIds = new StringBuilder("[");
-      for (int i = 0; i < rawFiducials.length; i++) {
-        fiducialIds.append(rawFiducials[i].id);
-        if (i < rawFiducials.length - 1) {
-          fiducialIds.append(", ");
-        }
-      }
-      fiducialIds.append("]");
-
-      return "PoseEstimate{"
-          + "pose="
-          + pose
-          + ", timestampSeconds="
-          + timestampSeconds
-          + ", latency="
-          + latency
-          + ", tagCount="
-          + tagCount
-          + ", tagSpan="
-          + tagSpan
-          + ", avgTagDist="
-          + avgTagDist
-          + ", avgTagArea="
-          + avgTagArea
-          + ", rawFiducials="
-          + fiducialIds
-          + ", isMegaTag2="
-          + isMegaTag2
-          + '}';
-    }
-  }
-
-  /**
-   * Encapsulates the state of an internal Limelight IMU. This includes gyro and accelerometer data.
-   */
-  public static class IMUData {
-    public double robotYaw;
-    public double roll;
-    public double pitch;
-    public double yaw;
-    public double gyroX;
-    public double gyroY;
-    public double gyroZ;
-    public double accelX;
-    public double accelY;
-    public double accelZ;
-
-    /** Initializes an "empty" IMUData object with default values */
-    public IMUData() {
-      this.robotYaw = 0;
-      this.roll = 0;
-      this.pitch = 0;
-      this.yaw = 0;
-      this.gyroX = 0;
-      this.gyroY = 0;
-      this.gyroZ = 0;
-      this.accelX = 0;
-      this.accelY = 0;
-      this.accelZ = 0;
-    }
-
-    /**
-     * Initializes an IMUData object with the provided IMU data array.
-     *
-     * @param robotYaw The yaw of the robot in degrees
-     * @param roll The roll of the robot in degrees
-     * @param pitch The pitch of the robot in degrees
-     * @param yaw The yaw of the IMU in degrees
-     * @param gyroX The x-axis gyro value in degrees per second
-     * @param gyroY The y-axis gyro value in degrees per second
-     * @param gyroZ The z-axis gyro value in degrees per second TODO: figure out what units the
-     *     accelerometer values are in (maybe Gs?)
-     * @param accelX The x-axis accelerometer value in meters per second squared
-     * @param accelY The y-axis accelerometer value in meters per second squared
-     * @param accelZ The z-axis accelerometer value in meters per second squared
-     */
-    public IMUData(
-        double robotYaw,
-        double roll,
-        double pitch,
-        double yaw,
-        double gyroX,
-        double gyroY,
-        double gyroZ,
-        double accelX,
-        double accelY,
-        double accelZ) {
-      this.robotYaw = robotYaw;
-      this.roll = roll;
-      this.pitch = pitch;
-      this.yaw = yaw;
-      this.gyroX = gyroX;
-      this.gyroY = gyroY;
-      this.gyroZ = gyroZ;
-      this.accelX = accelX;
-      this.accelY = accelY;
-      this.accelZ = accelZ;
-    }
-  }
 
   private static final String sanitizeName(String name) {
     if (name.equals("") || name.equals(null)) {
